Implement a basic Twirp RPC client generator for Ruby
Twitch recently released a gRPC alternative for Go and it looks
interesting. We're mostly a Ruby shop with some Go creeping in for a few
services. This puts the bones in place to start playing around with it
more.

This is very loosely based on the python generator that is part of
Twirp, though all of the code is original. I've chosen to use a Go
template because I've never used one before and wanted to give it a try.