Add a Ruby example to usage and mention the Haberdasher example
What was here left a lot to figure out so this should make getting
started easier.

One thing to keep in mind with the generated code is that protoc will
output two files: "#{file}_pb.rb" and "#{file}_twirp.rb". Both contain
code that cannot be autoloaded since the filename is not guaranteed to
match the module or class name. This means that both need to be required
manually and the protobuf definitions should be loaded before the client
file.